aggiunta parte delle lune

+++ Planetarium/src/it/unibs/eclisse/planetarium/Moon.java
@@ -2,7 +2,8 @@ package it.unibs.eclisse.planetarium;
 
 public class Moon extends CelestialBody{
 	
-	public Moon(String nome, double massa, int id, int x, int y) {
+	
+	public Moon(String nome, double massa, int id, double x, double y) {
 		
 		super(nome, massa, id, x, y);
 	}
